Single Page Application
Single-Page Applications (SPAs) are a kind of web applications that load
a single HTML page and dynamically update that page based ont he user
interaction. Single Page Applications are created with the help of
HTML5, CSS3 and JavaScript (mostly by using JavaScript frameworks or
plugins like jQuery, AngularJS, KnockoutJS etc.).

SPAs are responsive web applications in nature and because of that it
doesn't reloads the page on user interaction with the application. This
means that much of the work happens on the cleint side using JavaScript.